// client/src/styles/colors.ts
export interface IColors {
  $gray: string;
  $darkGray: string;
  $black: string;
  $red: string;
  $white: string;
  $orange: string;
  $yellow: string;
  $green: string;
  $blue: string;
  $absolutlyWhite: string;
  $pink: string;
  $purple: string;
}

export const colors: IColors = {
  $gray: '#C4C4C4',
  $darkGray: '#919191',
  $black: '#343434',
  $red: '#E80B0B',
  $white: '#FAFAFA',
  $orange: '#e8760b',
  $yellow: '#e8b80b',
  $green: '#3be80b',
  $blue: '#0be8e8',
  $absolutlyWhite: '#FFFFFF',
  $purple: '#633175',
  $pink: '#FF6565',
};